Grammar /Edit empty lines
