+++ gems.locked
@@ -1,12 +1,12 @@
 GIT
   remote: https://github.com/flaminestone/palladium.git
-  revision: cf36ed451f0d525f9053aed681f4f3e6831c59f2
+  revision: 48e301114606137ff21e4a39cbb4274c58c9acf8
   specs:
-    palladium (0.3.4)
+    palladium (0.3.3)
 
 GIT
   remote: https://github.com/onlyoffice-testing-robot/onlyoffice_documentserver_testing_framework
-  revision: 808a8d16fd5606e62657d7ef2f661c8ea2657935
+  revision: e740c6c08c99d6fbd1db5083f4c5a0d8920b39f2
   specs:
     onlyoffice_documentserver_testing_framework (0.0.1)
       onlyoffice_logger_helper (~> 1)
@@ -14,58 +14,52 @@ GIT
 
 GIT
   remote: https://github.com/onlyoffice-testing-robot/onlyoffice_tcm_helper
-  revision: a8de1f08d8de6923437f1bc1f7220b51bdf71e2c
+  revision: 68288baa932316abf9ed8fce2ce16e9b0bffb83b
   specs:
     onlyoffice_tcm_helper (0.1.1)
 
 GIT
   remote: https://github.com/onlyoffice-testing-robot/onlyoffice_webdriver_wrapper
-  revision: 73d77f942b58bbc1201458535b161a36ab831a5a
+  revision: b07fcf04304c66419471b9711d3a02f432f6c5cc
   specs:
-    onlyoffice_webdriver_wrapper (0.0.2)
+    onlyoffice_webdriver_wrapper (0.0.1)
       headless (~> 2)
       onlyoffice_file_helper (~> 0.1)
       onlyoffice_logger_helper (~> 1)
       onlyoffice_s3_wrapper (~> 0.1)
       page-object (~> 2.0.0)
-      selenium-webdriver (= 3.141.0)
+      selenium-webdriver (= 3.14.1)
       watir (~> 6.7.0)
 
 GEM
   remote: https://rubygems.org/
   specs:
     ast (2.4.0)
-    aws-eventstream (1.0.2)
-    aws-sdk (2.11.251)
-      aws-sdk-resources (= 2.11.251)
-    aws-sdk-core (2.11.251)
+    aws-sdk (2.11.158)
+      aws-sdk-resources (= 2.11.158)
+    aws-sdk-core (2.11.158)
       aws-sigv4 (~> 1.0)
       jmespath (~> 1.0)
-    aws-sdk-resources (2.11.251)
-      aws-sdk-core (= 2.11.251)
-    aws-sigv4 (1.1.0)
-      aws-eventstream (~> 1.0, >= 1.0.2)
+    aws-sdk-resources (2.11.158)
+      aws-sdk-core (= 2.11.158)
+    aws-sigv4 (1.0.3)
     childprocess (0.9.0)
       ffi (~> 1.0, >= 1.0.11)
-    concurrent-ruby (1.1.5)
+    concurrent-ruby (1.0.5)
     data_magic (1.2)
       faker (>= 1.1.2)
       yml_reader (>= 0.6)
     diff-lcs (1.3)
-    faker (1.9.3)
+    faker (1.9.1)
       i18n (>= 0.7)
-    ffi (1.10.0)
+    ffi (1.9.25)
     headless (2.3.1)
-    httparty (0.16.4)
-      mime-types (~> 3.0)
+    httparty (0.16.2)
       multi_xml (>= 0.5.2)
-    i18n (1.6.0)
+    i18n (1.1.1)
       concurrent-ruby (~> 1.0)
-    jaro_winkler (1.5.2)
+    jaro_winkler (1.5.1)
     jmespath (1.4.0)
-    mime-types (3.2.2)
-      mime-types-data (~> 3.2015)
-    mime-types-data (3.2019.0331)
     multi_xml (0.6.0)
     net-http-persistent (2.9.4)
     onlyoffice_file_helper (0.1.0)
@@ -82,10 +76,10 @@ GEM
       watir (~> 6.0)
     page_navigation (0.10)
       data_magic (>= 0.22)
-    parallel (1.17.0)
-    parser (2.6.2.0)
+    parallel (1.12.1)
+    parser (2.5.1.2)
       ast (~> 2.4.0)
-    psych (3.1.0)
+    powerpack (0.1.2)
     rainbow (3.0.0)
     rspec (3.8.0)
       rspec-core (~> 3.8.0)
@@ -100,20 +94,20 @@ GEM
       diff-lcs (>= 1.2.0, < 2.0)
       rspec-support (~> 3.8.0)
     rspec-support (3.8.0)
-    rubocop (0.66.0)
+    rubocop (0.60.0)
       jaro_winkler (~> 1.5.1)
       parallel (~> 1.10)
       parser (>= 2.5, != 2.5.1.1)
-      psych (>= 3.1.0)
+      powerpack (~> 0.1)
       rainbow (>= 2.2.2, < 4.0)
       ruby-progressbar (~> 1.7)
-      unicode-display_width (>= 1.4.0, < 1.6)
+      unicode-display_width (~> 1.4.0)
     ruby-progressbar (1.10.0)
     rubyzip (1.2.2)
-    selenium-webdriver (3.141.0)
+    selenium-webdriver (3.14.1)
       childprocess (~> 0.5)
       rubyzip (~> 1.2, >= 1.2.2)
-    unicode-display_width (1.5.0)
+    unicode-display_width (1.4.0)
     watir (6.7.3)
       selenium-webdriver (~> 3.4, >= 3.4.1)
     yml_reader (0.7)
@@ -133,4 +127,4 @@ DEPENDENCIES
   rubocop
 
 BUNDLED WITH
-   2.0.1
+   1.16.6
