+++ lib/testing_api_onlyoffice_com/test_instance/main_page/document_builder_api.rb
@@ -1,0 +1,2 @@
+require_relative 'document_builder_api/document_builder_introduction'
+require_relative 'document_builder_api/documen_builder_integrating'
