+++ lib/testing_api_onlyoffice_com/test_instance/main_page/document_builder_api/document_builder_api_common/left_side_navigation_builder.rb
@@ -1,21 +1,20 @@
-require 'httparty'
 require 'onlyoffice_file_helper'
 module TestingApiOnlyfficeCom
   # https://user-images.githubusercontent.com/18173785/37905775-9964ebb6-3108-11e8-8f98-480cbb1c2906.png
   # /docbuilder/basic
-  module LeftSideNavigation
+  module LeftSideNavigationBuilder
     include PageObject
 
     # actions
-    link(:introduction, xpath: '//a[@class="selected"]')
-    link(:getting_started, xpath: '//div[@class="layout-side"]//a[@href="/docbuilder/gettingstarted"]')
+    link(:introduction, xpath: '//*[contains(@class, "side-nav")]//a[@href="/docbuilder/basic"]')
+    link(:getting_started, xpath: '//*[contains(@href, "/docbuilder/gettingstarted")]')
     link(:integrating_document_builder, xpath: '//*[contains(@href, "/docbuilder/integratingdocumentbuilder")]')
     link(:integrating_document_builder_menu, xpath: '//li[@class="collapsable lastCollapsable"]/div[@class="hitarea collapsable-hitarea lastCollapsable-hitarea"]')
 
     link(:net_example, xpath: '//a[@href="/docbuilder/csharpexample"]')
     link(:nodejs_example, xpath: '//a[@href="/docbuilder/nodejsexample"]')
     link(:php_example, xpath: '//a[@href="/docbuilder/phpexample"]')
-    link(:ruby_example, xpath: '////a[@href="/docbuilder/rubyexample"]')
+    link(:ruby_example, xpath: '//a[@href="/docbuilder/rubyexample"]')
 
     link(:search_bar, xpath: '//div[@class="search-input"]')
     link(:search_button, xpath: '//div[@class="layout-side"]//a[@class="btn"]')
@@ -32,7 +31,7 @@ module TestingApiOnlyfficeCom
 
     def open_integrating_document_builder
       integrating_document_builder_element.click
-      TestingApiOnlyfficeCom::DocumentBuilderIntegrating.new(@instance)
+      DocumentBuilderIntegrating.new(@instance)
     end
   end
 end
