+++ lib/testing_api_onlyoffice_com/test_instance/main_page/document_builder_api/document_builder_introduction.rb
@@ -1,21 +1,12 @@
-require 'httparty'
 require 'onlyoffice_file_helper'
-require_relative 'left_side_navigation'
+require_relative 'document_builder_api_common/left_side_navigation_builder'
+require_relative 'document_builder_api_common/document_builder'
 module TestingApiOnlyfficeCom
   # https://user-images.githubusercontent.com/18173785/37905775-9964ebb6-3108-11e8-8f98-480cbb1c2906.png
   # /docbuilder/basic
   class DocumentBuilderIntroduction
-    include PageObject
-    include LeftSideNavigation
-
-    DEFAULT_BUILDER_FILE_NAME = 'SampleText.docx'.freeze
-
-
-    # actions
-    # introduction
-    link(:introduction, xpath: '//a[@class="selected"]')
-    link(:generate_document, xpath: '//*[@id="generateButton"]')
-
+    include LeftSideNavigationBuilder
+    include DocumentBuilder
 
     def initialize(instance)
       @instance = instance
@@ -24,20 +15,7 @@ module TestingApiOnlyfficeCom
     end
 
     def wait_to_load
-      @instance.webdriver.wait_until { generate_document_element.visible? }
+      @instance.webdriver.wait_until { button_generate_document_visible? }
     end
-
-    def builder_works?
-      generate_document_element.click
-      path_to_downloaded_file = @instance.webdriver.download_directory + '/' + DEFAULT_BUILDER_FILE_NAME
-      OnlyofficeFileHelper::FileHelper.wait_file_to_download(path_to_downloaded_file)
-      file_size = File.size(path_to_downloaded_file)
-      [file_size > 10_000, file_size]
-    end
-
-    def button_generate_document_visible?
-      generate_document_element.visible?
-    end
-
   end
 end
