+++ spec/site_api_spec.rb
@@ -20,19 +20,17 @@ describe 'site_api_tests' do
   end
 
   context '#document_builder' do
-    introduction_page = nil
-
     before :each do
-      introduction_page = @api_page.go_to_document_builder_introduction
+      @introduction_page = @api_page.go_to_document_builder_introduction
     end
 
     it '[API][DocumentBuilder] generate document works' do
-      result, file_size = introduction_page.builder_works?
+      result, file_size = @introduction_page.builder_works?
       expect(result).to be_truthy, "Page #{@instance.webdriver.driver.current_url}\n\nFile size: #{file_size}"
     end
 
     it '[API][DocumentBuilder] Integrating Document Builder download links shown and alive' do
-      integraing_page = introduction_page.open_integrating_document_builder
+      integraing_page = @introduction_page.open_integrating_document_builder
       result, failed = integraing_page.download_links_ok?
       expect(result).to be_truthy, "Page #{@instance.webdriver.driver.current_url}\n\nBad mojo with document builder links:\n #{failed}"
     end
